gui-apps/vicinae: re-restrict network-sandbox
Found some issues with arm64 due to native code being included with
`node_modules`, so we need to redo how we package `node_modules`.
Furthermore, system glaze won't work because it's too old, so we have to
let vicinae download it itself for now. Will take another pass at this
in the future.

Resolves https://github.com/jaredallard/overlay/issues/77